refactor: v3.6.0 maintainer CLI module consolidation
Decompose the post-TypeScript CLI entrypoint into flat command modules,
enable unused-code TypeScript flags, and align maintainer docs to src/dist
layout without changing public CLI behavior.
