+++ LICENSING.md
@@ -2,7 +2,7 @@
 
 | Area | License | Notes |
 | --- | --- | --- |
-| `bin/`, `scripts/`, package metadata  | MIT              | Original package code.                                                     |
+| `src/`, `dist/`, `scripts/`, package metadata | MIT         | Original package code.                                                     |
 | `docs/`, templates, examples, presets | MIT              | Original public content unless marked otherwise.                           |
 | `shared/references/tlc-baseline.md`   | CC-BY-4.0        | Adapted methodology with attribution in `NOTICE`. Upstream `tlc-spec-driven` pinned at `metadata.version: 3.3.0`. |
 | `shared/references/tdd-baseline.md`   | Attribution only | Inspired by `mattpocock/skills`; no separate source license claim is made. Upstream `tdd` skill carries no version of its own — pinned at repository release tag `v1.2.3`. |
