+++ presets/core.json
@@ -1,6 +1,6 @@
 {
   "name": "core",
-  "version": "3.5.0",
+  "version": "3.6.0",
   "description": "Safe baseline workflow.",
   "skills": [
     "saf-setup",
