+++ presets/full.json
@@ -1,6 +1,6 @@
 {
   "name": "full",
-  "version": "3.5.0",
+  "version": "3.6.0",
   "description": "All public skills.",
   "skills": [
     "saf-setup",
