+++ presets/github.json
@@ -1,6 +1,6 @@
 {
   "name": "github",
-  "version": "3.5.0",
+  "version": "3.6.0",
   "description": "Core and PR guidance for GitHub-aware agents.",
   "skills": [
     "saf-setup",
