+++ presets/local-files.json
@@ -1,6 +1,6 @@
 {
   "name": "local-files",
-  "version": "3.5.0",
+  "version": "3.6.0",
   "description": "Core workflow for local source files.",
   "skills": [
     "saf-setup",
