+++ presets/pr.json
@@ -1,6 +1,6 @@
 {
   "name": "pr",
-  "version": "3.5.0",
+  "version": "3.6.0",
   "description": "Read-first PR workflow.",
   "skills": ["saf-route", "saf-create-pr", "saf-review-pr", "saf-fix-pr"],
   "shared": true,
