+++ skills/saf-check-task/SKILL.md
@@ -2,7 +2,7 @@
 name: saf-check-task
 description: Independently check one implemented SDD task against its acceptance criteria and configured gates before handoff. Use for a task-scoped readiness check, not feature-wide validation or code changes.
 metadata:
-  version: 3.5.0
+  version: 3.6.0
   pack: core
 extends: saf-implement
 requires: [config, task-evidence]
