+++ skills/saf-review-pr/SKILL.md
@@ -2,7 +2,7 @@
 name: saf-review-pr
 description: Review one task-scoped pull request against its SDD, diff, and configured checks. Use for an evidence-based PR review; not for fixing findings or mutating PR metadata.
 metadata:
-  version: 3.5.0
+  version: 3.6.0
   pack: pr
 extends: saf-create-pr
 requires: [config, pr-reference]
