+++ skills/saf-route/SKILL.md
@@ -2,7 +2,7 @@
 name: saf-route
 description: Recommend the next local SDD skill without changing files. Use when a user needs help choosing a safe workflow step or resolving prerequisites.
 metadata:
-  version: 3.5.0
+  version: 3.6.0
   pack: core
 extends: null
 requires: [config]
