+++ skills/saf-setup/SKILL.md
@@ -2,7 +2,7 @@
 name: saf-setup
 description: Initialize the public SDD Agentic Flow structure in a repository. Use when a user asks to set up, bootstrap, or configure this SDD workflow; create only the requested repository-local planning files and start from .sdd-agentic-flow/config.yml.
 metadata:
-  version: 3.5.0
+  version: 3.6.0
   pack: core
 extends: null
 requires: [config]
